Introduce proto-format flag to store data in proto
We store data in json because of readability. But having this format limit us to the version of api bundled into the binary. With optional moving to proto format we can use the same version of api if it will not include changes in AuthContainer logic and will not introduce global changes in how to handle objects.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>